requirements.txt
streamlit==1.44.1
reportlab==4.3.1
pillow==11.1.0
PyPDF2==3.0.1